Add linking to libdl for coreclr on macOS
Currently we rely on the fact that dlopen is implemented in libSystem.dylib.
But there is a libdl.dylib on macOS too. That is just a symbolic link to
the libSystem.dylib, but it can change in the future. So adding libdl
to the set of libraries we link to is a right thing.