Fix coredistools.dll missing from r2rdump directory
This change fixes a problem when coredistools.dll was missing from the
R2RDump binary directory after build and it needed to be copied there
manually.
The real issue is likely hidden in the Microsoft.NETCore.CoreDisTools
package, but this fix uses the same trick as stress_dependecies.csproj
to find the runtime.$RID.Microsoft.NETCore.CoreDisTools package and copy
the binary from there.