Fix LLD support
I have found that when I've made a change to use the lld linker in July,
I've missed to push part of the change that was supposed to go to Arcade.

I am checking it to the runtime repo to ensure that it is fixed for the
RC2 release. I will make the same change in the Arcade repo.