Remove useless PAL_DisplayDialog
The code for the dialog was pulling in about 90 shared libraries on
macOS, total RSS of those being 512kB. This dialog is only shown in
debug builds and only when a special env variable was set. Noone used
that ever since the .NET Core 1.0 and yet we pay this unnecessary price.

This change removes the function and the dependency on unneeded
libraries.