Fix Linux arm64 docker hang with unhandled exception
When unhandled exception occurs in a docker container and the failing
process was started using the entry point of the image, the process
hangs after reporting the unhandled exception forever.

The problem is caused by incorrect `signalRestarts` flag passed to the
`invoke_previous_action` by the `sigabrt_handler`. Instead of aborting
the process, it returns back to the code that issued the SIGABRT (the C
library `abort` implementation). On arm64, when that signal doesn't
abort the process, it hits a `brk` instruction. That results in SIGTRAP
signal that ultimately calls `IsSafeToHandleHardwareException` that
returns `false` and the SIGTRAP handler returns back, hits the `brk`
again and it keeps going on forever.

The fix is to pass `signalRestarts` set to `false` to the
`invoke_previous_action` in the `sigabrt_handler`. Then the
`invoke_previous_action` calls `PROCAbort` that disables all signal
handlers and calls `abort` again.