Fix unhandled exception with no managed frames
There is a case when there are no managed frames on the stack nor a
native unhandled exception trap in the native code. For example when a
custom host invokes a delegate and it ends up jitting an pinvoke code
and calls NDirectImportWorker that throws because of a problem with the
pinvoke.

The exception handling code in DispatchManagedException was expecting
that there is either an unhandled managed exception trap installed or
managed code on the stack that would both "handle" the exception
properly. But in the case of a delegate invoked by a custom or a thread
with purely native code it is not the case and the runtime ends up
crashing with PAL_SEHException leaking out of the runtime.

This change fixes it by adding explicit call to the unhandled exception
filter to the DispatchManagedException when it doesn't find any managed
frames on the stack.