Further improve the software EH performance
This change gets rid of the last remaining call to unwind native frame
in IL_Throw and IL_Rethrow. It uses asm helper to capture the context
at the call site of IL_Throw and IL_Rethrow. The performance gain on
Windows is 5% in single threaded and 20% in multi threaded cases of a
test that throws software exception over 10 managed frames.